Reverse the order of events on homepage.
This isn't just a hack, though it looks like it. Maybe I've just had one
too many. But we return events sorted by date. For the events page, that
works out such that the farthest future event is at the top, somewhere in
the middle is present day, and the farthest past event is at the bottom.
This is the opposite of what we want for the homepage, which only shows
future events. So just reverse the order for the homepage.

Wow that took way too long to figure out.